requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
python-multipart
easyocr
pandas
faiss-cpu
numpy
scipy
opencv-python
scikit-image
scikit-learn